\name{oak.rads}
\docType{data}
\alias{oak.rads}
\title{101-locus RAD dataset, in \code{pyRAD.loci} format}
\description{This dataset includes 101 RAD loci from a recent phylogenetic study of oaks (Hipp et al. in review).}
\usage{oak.rads}
\format{An object of class \code{pyRAD.loci}}
\references{Hipp A.L., Eaton D.A.R., Cavender-Bares J., Fitzek E., Nipper R. & Manos P.S. (Accepted pending revision). A framework phylogeny of the American oak clade based on sequenced RAD data. PLoS ONE.}
\keyword{datasets}
