\name{oak.tree}
\docType{data}
\alias{oak.tree}
\title{28-tip oak phylogeny, in \code{phylo} format}
\description{This dataset is the 28-tip oak framework oak phylogeny 
			 published in Hipp et al. (in revision). Tips in this 
			 phylogeny match up with the tips in the oak RAD dataset
			 provided with this package.}
\usage{oak.tree}
\format{An object of class \code{phylo}}
\references{Hipp A.L., Eaton D.A.R., Cavender-Bares J., Fitzek E., Nipper R. & Manos P.S. (Accepted pending revision). A framework phylogeny of the American oak clade based on sequenced RAD data. PLoS ONE.}
\seealso{
  \code{\link{oak.rads}}
  }
\keyword{datasets}