\name{RADami-package}
\alias{RADami-package}
\alias{RADami}
\docType{package}
\title{Phylogenetic Analysis of RADseq Data}
\description{This package implements import, export, manipulation, visualization, and downstream
             (post-clustering) analysis of RADseq data, integrating with the pyRAD package by Deren Eaton.
			 It begins with output from the pyRAD package by Deren Eataon (http://dereneaton.com/software; Eaton and Ree, 2013) and
			 implements many analyses conducted in a recent phylogenetic analysis of oaks (Hipp et al., accepted).}
\details{
\tabular{ll}{
Package: \tab RADami\cr
Type: \tab Package\cr
Version: \tab 1.1-2\cr
Date: \tab 2017-02-10\cr
License: \tab GPL-3\cr
}
}
\author{Andrew Hipp

Maintainer: Andrew Hipp <ahipp@mortonarb.org>
}
\references{
Eaton D.A.R. & Ree R.H. (2013). Inferring Phylogeny and Introgression using RADseq Data: An Example from Flowering Plants (Pedicularis: Orobanchaceae). Systematic Biolology 62: 689-706.

Hipp A.L., Eaton D.A.R., Cavender-Bares J., Fitzek E., Nipper R. & Manos P.S. 2014. A framework phylogeny of the American oak clade based on sequenced RAD data. PLoS ONE 9: e93975.
http://www.dereneaton.com/software/pyrad/
}
\keyword{ package }
